회원 가입 체육관 리스트 개편 / 전화번호 스타일 파셜화 / 회원가입시 추천인 코드 입력하는 기능 일부 구현

+++ db/migrate/20200518054234_remove_item_price_from_line_items.rb
@@ -1,6 +1,8 @@
 class RemoveItemPriceFromLineItems < ActiveRecord::Migration[6.0]
   def change
-    remove_column :line_items, :item_id
-    remove_column :line_items, :price
+    if table_exists?(:line_items)
+      remove_column :line_items, :item_id
+      remove_column :line_items, :price
+    end
   end
 end
